new bransch commit

+++ MyZoo/DAL/DataAccessZoo.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Data.Entity;
 using System.Data.Entity.Core;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using Microsoft.SqlServer.Server;
 using MyZoo.DataContext;
 using MyZoo.Model;
@@ -30,6 +32,9 @@ namespace MyZoo.DAL
                         from animal in db.Animals
                         where animal.Species.Name.ToLower()
                             .Contains(search.SpeciesSearch.ToLower())
+                        join parent in db.ParentCouples
+                            on animal.ParentCouple.ParentCoupleId equals parent.ParentCoupleId into ap
+                        from p in ap.DefaultIfEmpty()
                         select new AnimalDetailed()
                         {
                             AnimalId = animal.AnimalId,
@@ -42,8 +47,8 @@ namespace MyZoo.DAL
                             Species = animal.Species.Name,
                             Type = animal.Species.Type.Name,
                             WeightInKilogram = animal.Weight,
-                            ChildList = animal.ParentCouple.Animals.Select(a => a.Name).ToList()
                         };
+
                     list = new BindingList<AnimalDetailed>(query.ToList());
                 }
                 else
@@ -65,7 +70,6 @@ namespace MyZoo.DAL
                             Species = animal.Species.Name,
                             Type = animal.Species.Type.Name,
                             WeightInKilogram = animal.Weight,
-                            ChildList = p.Animals.Select(a => a.Name).ToList()
 
                         };
 
@@ -81,8 +85,34 @@ namespace MyZoo.DAL
         {
             using (var db = new ZooDBContext())
             {
+                var animalQuery = from animal in db.Animals
+                                  where animal.AnimalId == animalId
+                                  select animal;
 
-                db.Animals.Remove(db.Animals.Find(animalId));
+                var mother = animalQuery.Select(a => a.ParentCouple.Mother);
+
+                var father = animalQuery.Select(f => f.ParentCouple.Father);
+
+                var list = animalQuery;
+                //list.Select(p => p.ParentCouple).Select(f => f.Father) = father
+
+                if (db.Mothers.Any(a => a.Animal.AnimalId == animalId))
+                {
+                    var couple = db.ParentCouples.Find(db.Mothers.FirstOrDefault(m => m.Animal.AnimalId == animalId));
+                    var mothr = db.Animals.Find(animalId);
+                    db.Animals.Remove(db.Animals.Find(animalId));
+                }
+                if (db.Fathers.Any(a => a.Animal.AnimalId == animalId))
+                {
+                    var couple = db.ParentCouples.Find(db.Fathers.FirstOrDefault(m => m.Animal.AnimalId == animalId));
+                    couple.Father = null;
+                    db.Animals.Remove(db.Animals.Find(animalId));
+                }
+                else
+                {
+                    db.Animals.Remove(db.Animals.Find(animalId));
+                }
+
                 db.SaveChanges();
             }
         }
